Refactored StaffAutoCompleteField, created it as reusable components and  moved TaskContext to dashboard scope.
