+++ nextjs/src/app/dashboard/Provider.tsx
@@ -6,6 +6,7 @@ import CustomTheme from "../../theme";
 import { StaffContextProvider } from "@/context/StaffContext";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { LocalizationProvider } from "@mui/x-date-pickers";
+import { TasksContextProvider } from "@/context/TasksContext";
 
 interface Props {
   children: React.ReactNode;
@@ -17,7 +18,9 @@ function Provider({ children }: Props) {
   return (
     <LocalizationProvider dateAdapter={AdapterDayjs}>
       <ThemeProvider theme={forcedLight}>
-        <StaffContextProvider>{children}</StaffContextProvider>
+        <StaffContextProvider>
+          <TasksContextProvider>{children}</TasksContextProvider>
+        </StaffContextProvider>
       </ThemeProvider>
     </LocalizationProvider>
   );
