+++ nextjs/src/app/dashboard/tasks/layout.tsx
@@ -1,4 +1,4 @@
-import Provider from "./Provider";
+import Provider from "@/app/dashboard/Provider";
 import React, { useState, useEffect } from "react";
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
